+++ pkg/storage/stores/tsdb/multi_file_index_test.go
@@ -8,10 +8,10 @@ import (
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/stretchr/testify/require"
 
-	"github.com/grafana/loki/pkg/storage/tsdb/index"
+	"github.com/grafana/loki/pkg/storage/stores/tsdb/index"
 )
 
-func TestSingleIdx(t *testing.T) {
+func TestMultiIndex(t *testing.T) {
 	cases := []LoadableSeries{
 		{
 			Labels: mustParseLabels(`{foo="bar"}`),
@@ -44,21 +44,31 @@ func TestSingleIdx(t *testing.T) {
 			},
 		},
 		{
-			Labels: mustParseLabels(`{foo="bard", bazz="bozz", bonk="borb"}`),
+			// should be excluded due to bounds checking
+			Labels: mustParseLabels(`{foo="bar", bazz="bozz", bonk="borb"}`),
 			Chunks: []index.ChunkMeta{
 				{
-					MinTime:  1,
-					MaxTime:  7,
+					MinTime:  8,
+					MaxTime:  9,
 					Checksum: 4,
 				},
 			},
 		},
 	}
 
-	idx := BuildIndex(t, t.TempDir(), "fake", cases)
+	// group 5 indices together, all with duplicate data
+	n := 5
+	var indices []*TSDBIndex
+	dir := t.TempDir()
+	for i := 0; i < n; i++ {
+		indices = append(indices, BuildIndex(t, dir, "fake", cases))
+	}
+
+	idx, err := NewMultiIndex(indices...)
+	require.Nil(t, err)
 
 	t.Run("GetChunkRefs", func(t *testing.T) {
-		refs, err := idx.GetChunkRefs(context.Background(), "fake", 1, 5, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
+		refs, err := idx.GetChunkRefs(context.Background(), "fake", 2, 5, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
 		require.Nil(t, err)
 
 		expected := []ChunkRef{
@@ -94,79 +104,54 @@ func TestSingleIdx(t *testing.T) {
 		require.Equal(t, expected, refs)
 	})
 
-	t.Run("GetChunkRefsSharded", func(t *testing.T) {
-		shard := index.ShardAnnotation{
-			Shard: 1,
-			Of:    2,
-		}
-		shardedRefs, err := idx.GetChunkRefs(context.Background(), "fake", 1, 5, nil, &shard, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
-
-		require.Nil(t, err)
-
-		require.Equal(t, []ChunkRef{{
-			User:        "fake",
-			Fingerprint: model.Fingerprint(mustParseLabels(`{foo="bar", bazz="buzz"}`).Hash()),
-			Start:       1,
-			End:         10,
-			Checksum:    3,
-		}}, shardedRefs)
-
-	})
-
 	t.Run("Series", func(t *testing.T) {
-		xs, err := idx.Series(context.Background(), "fake", 8, 9, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
+		xs, err := idx.Series(context.Background(), "fake", 2, 5, nil, nil, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
 		require.Nil(t, err)
-
-		expected := []Series{
-			{
-				Labels:      mustParseLabels(`{foo="bar", bazz="buzz"}`),
-				Fingerprint: model.Fingerprint(mustParseLabels(`{foo="bar", bazz="buzz"}`).Hash()),
-			},
-		}
-		require.Equal(t, expected, xs)
-	})
-
-	t.Run("SeriesSharded", func(t *testing.T) {
-		shard := index.ShardAnnotation{
-			Shard: 0,
-			Of:    2,
-		}
-
-		xs, err := idx.Series(context.Background(), "fake", 0, 10, nil, &shard, labels.MustNewMatcher(labels.MatchEqual, "foo", "bar"))
-		require.Nil(t, err)
-
 		expected := []Series{
 			{
 				Labels:      mustParseLabels(`{foo="bar"}`),
 				Fingerprint: model.Fingerprint(mustParseLabels(`{foo="bar"}`).Hash()),
 			},
+			{
+				Labels:      mustParseLabels(`{foo="bar", bazz="buzz"}`),
+				Fingerprint: model.Fingerprint(mustParseLabels(`{foo="bar", bazz="buzz"}`).Hash()),
+			},
 		}
+
 		require.Equal(t, expected, xs)
 	})
 
 	t.Run("LabelNames", func(t *testing.T) {
 		// request data at the end of the tsdb range, but it should return all labels present
-		ls, err := idx.LabelNames(context.Background(), "fake", 9, 10)
+		xs, err := idx.LabelNames(context.Background(), "fake", 8, 10)
 		require.Nil(t, err)
-		require.Equal(t, []string{"bazz", "bonk", "foo"}, ls)
+		expected := []string{"bazz", "bonk", "foo"}
+
+		require.Equal(t, expected, xs)
 	})
 
 	t.Run("LabelNamesWithMatchers", func(t *testing.T) {
 		// request data at the end of the tsdb range, but it should return all labels present
-		ls, err := idx.LabelNames(context.Background(), "fake", 9, 10, labels.MustNewMatcher(labels.MatchEqual, "bazz", "buzz"))
+		xs, err := idx.LabelNames(context.Background(), "fake", 8, 10, labels.MustNewMatcher(labels.MatchEqual, "bazz", "buzz"))
 		require.Nil(t, err)
-		require.Equal(t, []string{"bazz", "foo"}, ls)
+		expected := []string{"bazz", "foo"}
+
+		require.Equal(t, expected, xs)
 	})
 
 	t.Run("LabelValues", func(t *testing.T) {
-		vs, err := idx.LabelValues(context.Background(), "fake", 9, 10, "foo")
+		xs, err := idx.LabelValues(context.Background(), "fake", 1, 2, "bazz")
 		require.Nil(t, err)
-		require.Equal(t, []string{"bar", "bard"}, vs)
+		expected := []string{"bozz", "buzz"}
+
+		require.Equal(t, expected, xs)
 	})
 
 	t.Run("LabelValuesWithMatchers", func(t *testing.T) {
-		vs, err := idx.LabelValues(context.Background(), "fake", 9, 10, "foo", labels.MustNewMatcher(labels.MatchEqual, "bazz", "buzz"))
+		xs, err := idx.LabelValues(context.Background(), "fake", 1, 2, "bazz", labels.MustNewMatcher(labels.MatchEqual, "bonk", "borb"))
 		require.Nil(t, err)
-		require.Equal(t, []string{"bar"}, vs)
+		expected := []string{"bozz"}
+
+		require.Equal(t, expected, xs)
 	})
 }